refactor: extract RecordSets as a package-private collaborator
The merged/pass-through record collections and the per-type identity
functions roughly doubled OneshotRecordParser; move them to their own
package-private class so the parser keeps to parsing and assembly wiring.
No behavior change.
